Official Rho-Theta code commit - March 18, 2015
1) Added gravity terms to horizontal wind updates in
HorizontalDynamicsFEM.cpp (probably needs to be done in the DG code too)

2) Implements the SLEVE coordinate in the vertical in
GridPatchCartesianGLL.cpp (better for high order vertical elements)

3) Broken code for output of energy in cartesian cases in GridPatch.cpp

4) Changed the initialization templates in TempestInitialize.h to
include a reference latitude and maximum topography height... this is
used in the SLEVE scale height and setting a beta plane approximation
coriolis force. Both of these are done in GridPatchCartesianGLL.cpp

5) Vertical dynamics should be unchanged but it needs updating to
include vertical gradient terms that update the horizontal wind

6) Added a broken version of the 3D baroclinic instability case in a
channel. It is broken because no-flux boundaries in beta direction are
not implemented

Enjoy! JEG